Workstealing: Use upper 16 PRNG bits
Because of its parameters, the pseudo random number generator provides a good
periodicity. However, this is only valid for the upper bits. The modulus required
by the selection of a target work deque during work stealing discards the upper bits
and thus only keeps the lower bits. The result is a very short period that might
be smaller than the total number of workers. This leads to poor performance if
work deques are imbalanced.

In order to improve the period of the target queue random number generator,
only the upper bits are used by shifting the PRNG's value by 16 prior to
the modulus operation.