Use OpenStream GCC to compile and link libwstream_df.so
Since the addition of the MPSC FIFO a C11 compliant compiler is required for
libwstream_df.so. The modified GCC version for OpenStream is guranteed to
support the necessary C11 features, so using it in the Makefile takes the
burden of compiling an extra GCC or setting the CC environment variable on
systems where no C11 compliant compiler is available off the shoulder of the
user.