Make 2-pass tcreate
During the first pass only allocate a frame for metadata (task private data, dependences, etc.).
Then, prematch the new frame with its producers and consumers in order to determine which NUMA
node provides / consumes most of the task's data.

During the second pass, allocate the frame on the node with the strongest dependences copy the
metadata and perform definite matching with producers and consumers.