BUGFIX: Make sure a task does not get ready until all dependences have been considered
Variadic views might have 0 dependences. Such dependences are not
included in the reference counter of the task to be created. If such
they are resolved towards the end of task creation, the task might
become ready too soon: another worker might execute it before the last
call to resolve_dependence and the creating worker continues with
invalid reference.

Adding 1 to the synchronization counter and decrementing it at the
very end of task creation solves this problem.