BUGFIX: Do not interpret arrays of streams of size 1 as single streams in stream destructor
The distinction between arrays of streams and single streams in
wstream_df_stream_dtor is based on the number of streams. However, for
both arrays of size 1 and single streams the number of streams passed
to the destructor is 1. This causes the first element of an array of
size 1 to be considered as a stream instance and not as a pointer to a
stream instance.

This patch splits the destructor into an array destructor and a
destructor for single streams.