Add support for deferred allocation
Buffers of input views get allocated at task creation, which might
lead to a huge memory footprint of the application. However, if for an
input view there is only a single producer, the view buffer can be
allocated when the producer becomes ready. This feature is enabled if
DEFERRED_ALLOC is set in the configuration header [enabled in all
default configurations].