BUGFIX: Do not call wstream_df_num_cores after pinning worker threads to CPUs
wstream_df_num_cores determines the number of cores by counting the
number of bits set in the bitmask that defined on which CPUs a thread
can be scheduled. When this function is called from a thread pinned to
a CPU the return value of wstream_df_num_cores is thus incorrect. This
patch introduces a new function wstream_df_num_cores_cached that
returns a cached value.