Add options --openstream-only and --body-funcs-only to seidel.py
When using --openstream-only the generator produces code that only
consists of the OpenStream part of a parallel benchmark. The option
--body-funcs-only does the exact opposite and only generates code that
is not OpenStream-specific. This allows for the generation of distinct
source files for the OpenStream part and the code used by the task
bodies of the main computational tasks. The task bodies can thus be
compiled using a different compiler.