Stream_df_cholesky_peek: Guarantee data is copied back to shared memory
Previously, the control program could proceed to halt the timer and
enter the verification stage w/o guaranteeing the matrix was fully
unloaded from the streams to shared memory. This caused verification
to fail on single-threaded runs. This was not a problem in
multi-threaded runs, presumably because the other threads would be free
to finish these copies before the copy of the last block (which the
control programs always waited for) happened.